feat(parservX): add ast.Walk checking parent scope
We need a variation of the ast.Walk function which
keeps track of usage of parent scope sign ^ vs.
the usage of filters or projections.

This is exclusively used to walk the body of
custom defined functions, as we don't allow
the usage of parent scope which reaches out of the
function.

Co-authored-by: Sindre Gulseth <sgulseth@gmail.com>